+++ pysui/sui/sui_types/collections.py
@@ -113,6 +113,11 @@ class SuiMap(SuiCollection):
         """Alias for transactions."""
         return self.map
 
+    @property
+    def filter(self) -> dict[str, Any]:
+        """Alias for transactions."""
+        return self.map
+
 
 class EventID(SuiMap):
     """Event ID specifier."""